ML_Service/requirements.txt
fastapi
pydantic
uvicorn
pandas
numpy
scikit-learn
xgboost
lightgbm
torch
torchvision
Pillow
python-multipart
